+++ examples/1d_evaluate.py
@@ -31,7 +31,7 @@ def evaluate(filename):
 
     lrm = PRModel(1, page_size)
     prm = PRModel(2, page_size)
-    sgm = StagedModel(['fcn', 'fcn', 'lr'], [1, 20, 100000], page_size)
+    sgm = StagedModel(['fcn', 'fcn', 'lr'], [1, 20, 10000], page_size)
     models = [sgm, btm, fcn, lrm, prm]
     ptq = PointQuery(models)
     build_times = ptq.build(data, ratio)